requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain-huggingface
langchain_experimental
google-generativeai
groq
pinecone-client
# -e .